Remove Swift Testing.
