--- Tests/VecLabTests/SwiftTests/Array/ArrayTests2.swift
@@ -1,8 +1,0 @@
-//
-//  ArrayTests.swift
-//  VecLab
-//
-//  Created by Marcus Painter on 22/02/2025.
-//
-
-class ArrayTests2 { }
