--- Tests/VecLabTests/SwiftTests/Logical/LogicalTests.swift
@@ -1,8 +1,0 @@
-//
-//  LogicalTests.swift
-//  VecLab
-//
-//  Created by Marcus Painter on 23/02/2025.
-//
-
-class LogicalTests { }
